omit commits already present on other branches
If
 1 branch_a is branched from master
 2 a commit with a bug number is made and pushed to remote_host master
 3 branch_a merges in changes from master
 4 branch_a is pushed to remote_host branch_a
then gitzilla will post to bugzilla in step 2, and then post the
same bug again in step 4, because the commit made in master is
new to branch_a.

With this change, I use the strategy that the post-receive-email
hook included with git uses to exclude commits that are already
present on other branches.